refactor fill_utils.py
