+++ keras_cv/layers/preprocessing/cut_mix.py
@@ -126,8 +126,8 @@ class CutMix(layers.Layer):
                 images,
                 random_center_width,
                 random_center_height,
-                cut_width // 2,
-                cut_height // 2,
+                cut_height,
+                cut_width,
                 tf.gather(images, permutation_order),
             ),
             fn_output_signature=tf.TensorSpec.from_tensor(images[0]),
