+++ keras_cv/layers/preprocessing/random_cutout.py
@@ -151,16 +151,14 @@ class RandomCutout(layers.Layer):
         center_x, center_y = self._compute_rectangle_position(inputs)
         rectangle_height, rectangle_width = self._compute_rectangle_size(inputs)
         rectangle_fill = self._compute_rectangle_fill(inputs)
-        half_height = tf.cast(tf.math.ceil(rectangle_height / 2), tf.int32)
-        half_width = tf.cast(tf.math.ceil(rectangle_width / 2), tf.int32)
         inputs = tf.map_fn(
             lambda x: fill_utils.fill_rectangle(*x),
             (
                 inputs,
-                center_y,
                 center_x,
-                half_width,
-                half_height,
+                center_y,
+                rectangle_height,
+                rectangle_width,
                 rectangle_fill,
             ),
             fn_output_signature=tf.TensorSpec.from_tensor(inputs[0]),
@@ -177,14 +175,14 @@ class RandomCutout(layers.Layer):
         center_x = tf.random.uniform(
             shape=[batch_size],
             minval=0,
-            maxval=image_height,
+            maxval=image_width,
             dtype=tf.int32,
             seed=self.seed,
         )
         center_y = tf.random.uniform(
             shape=[batch_size],
             minval=0,
-            maxval=image_width,
+            maxval=image_height,
             dtype=tf.int32,
             seed=self.seed,
         )
