+++ keras_cv/utils/fill_utils.py
@@ -14,17 +14,86 @@
 import tensorflow as tf
 
 
-def fill_rectangle(
-    image, center_width, center_height, half_width, half_height, fill=None
-):
+def rectangle_masks(images, x0, y0, x1, y1):
+    """positional masks of rectangles in images"""
+    images_shape = tf.shape(images)
+    batch_size, image_height, image_width = (
+        images_shape[0],
+        images_shape[1],
+        images_shape[2],
+    )
+
+    # add broadcasting axes
+    x0 = x0[:, tf.newaxis, tf.newaxis]
+    y0 = y0[:, tf.newaxis, tf.newaxis]
+    x1 = x1[:, tf.newaxis, tf.newaxis]
+    y1 = y1[:, tf.newaxis, tf.newaxis]
+
+    # repeat for height and width
+    x0_rep = tf.repeat(x0, image_height, axis=1)
+    y0_rep = tf.repeat(y0, image_width, axis=2)
+    x1_rep = tf.repeat(x1, image_height, axis=1)
+    y1_rep = tf.repeat(y1, image_width, axis=2)
+
+    # range grid
+    range_row = tf.range(0, image_height)
+    range_col = tf.range(0, image_width)
+    range_row = tf.repeat(range_row[tf.newaxis, :, tf.newaxis], batch_size, 0)
+    range_col = tf.repeat(range_col[tf.newaxis, tf.newaxis, :], batch_size, 0)
+
+    # boolean masks
+    mask_x0 = tf.less_equal(x0_rep, range_col)
+    mask_y0 = tf.less_equal(y0_rep, range_row)
+    mask_x1 = tf.less(range_col, x1_rep)
+    mask_y1 = tf.less(range_row, y1_rep)
+
+    masks = mask_x0 & mask_y0 & mask_x1 & mask_y1
+
+    return masks
+
+
+def batch_fill_rectangle(images, center_x, center_y, height, width, fill):
+    """Fill a rectangle in a given image using the value provided in replace.
+
+    Args:
+        images: the starting image to fill the rectangle on.
+        center_x: the X center of the rectangle to fill
+        center_y: the Y center of the rectangle to fill
+        height: the height of the resulting rectangle
+        width: the width of the resulting rectangle
+        fill: A tensor with same shape as image. Values at rectangle
+         position are used as fill.
+    Returns:
+        image: the modified image with the chosen rectangle filled.
+    """
+    images_shape = tf.shape(images)
+    image_height = images_shape[1]
+    image_width = images_shape[2]
+
+    half_height = tf.cast(tf.math.ceil(height / 2), tf.int32)
+    half_width = tf.cast(tf.math.ceil(width / 2), tf.int32)
+
+    x0 = tf.maximum(0, center_x - half_width)
+    x1 = tf.minimum(image_width, center_x + half_width)
+    y0 = tf.maximum(0, center_y - half_height)
+    y1 = tf.minimum(image_height, center_y + half_height)
+
+    is_patch_mask = rectangle_masks(images, x0, y0, x1, y1)
+    is_patch_mask = tf.expand_dims(is_patch_mask, -1)
+
+    images = tf.where(tf.equal(is_patch_mask, True), fill, images)
+    return images
+
+
+def fill_rectangle(image, center_x, center_y, height, width, fill=None):
     """Fill a rectangle in a given image using the value provided in replace.
 
     Args:
         image: the starting image to fill the rectangle on.
-        center_width: the X center of the rectangle to fill
-        center_height: the Y center of the rectangle to fill
-        half_width: 1/2 the width of the resulting rectangle
-        half_height: 1/2 the height of the resulting rectangle
+        center_x: the X center of the rectangle to fill
+        center_y: the Y center of the rectangle to fill
+        height: the height of the resulting rectangle
+        width: the width of the resulting rectangle
         fill: A tensor with same shape as image. Values at rectangle
          position are used as fill.
     Returns:
@@ -34,10 +103,13 @@ def fill_rectangle(
     image_height = image_shape[0]
     image_width = image_shape[1]
 
-    lower_pad = tf.maximum(0, center_height - half_height)
-    upper_pad = tf.maximum(0, image_height - center_height - half_height)
-    left_pad = tf.maximum(0, center_width - half_width)
-    right_pad = tf.maximum(0, image_width - center_width - half_width)
+    half_height = tf.cast(tf.math.ceil(height / 2), tf.int32)
+    half_width = tf.cast(tf.math.ceil(width / 2), tf.int32)
+
+    lower_pad = tf.maximum(0, center_y - half_height)
+    upper_pad = tf.maximum(0, image_height - center_y - half_height)
+    left_pad = tf.maximum(0, center_x - half_width)
+    right_pad = tf.maximum(0, image_width - center_x - half_width)
 
     shape = [
         image_height - (lower_pad + upper_pad),
